Allow partial loading of records

+++ crates/integration-tests/src/load_data.rs
@@ -1,5 +1,6 @@
 #[cfg(test)]
 mod tests {
+    use std::env;
     use assert_cmd::prelude::*;
     use predicates::prelude::*;
     use std::io::Write;
@@ -13,12 +14,19 @@ mod tests {
 
     use minerva::schema::create_schema;
     use minerva::trend_store::{create_partitions_for_timestamp, AddTrendStore, TrendStore};
+    use rust_decimal::Decimal;
+    use rust_decimal_macros::dec;
 
     const TEST_CSV_DATA: &str = r###"
 node,timestamp,outside_temp,inside_temp,power_kwh,freq_power
 hillside14,2023-03-25T14:00:00Z,14.4,32.4,55.8,212.4
 hillside15,2023-03-25T14:00:00Z,14.5,32.5,55.9,212.5
-    "###;
+"###;
+
+    const TEST_CSV_DATA_UPDATE_PARTIAL: &str = r###"
+node,timestamp,power_kwh,freq_power
+hillside15,2023-03-25T14:00:00Z,55.9,200.0
+"###;
 
     const TREND_STORE_DEFINITION: &str = r###"
     title: Raw node data
@@ -52,6 +60,7 @@ hillside15,2023-03-25T14:00:00Z,14.5,32.5,55.9,212.5
     #[cfg(test)]
     #[tokio::test]
     async fn load_data() -> Result<(), Box<dyn std::error::Error>> {
+        let keep_database = env::var("DROP_DATABASE").unwrap_or(String::from("1")).eq("0");
         let data_source_name = "hub";
         let database_name = generate_name();
         let db_config = get_db_config()?;
@@ -95,11 +104,13 @@ hillside15,2023-03-25T14:00:00Z,14.5,32.5,55.9,212.5
             .success()
             .stdout(predicate::str::contains("Job ID"));
 
-        let mut client = connect_to_db(&db_config).await?;
+        if !keep_database {
+            let mut client = connect_to_db(&db_config).await?;
 
-        drop_database(&mut client, &database_name).await?;
+            drop_database(&mut client, &database_name).await?;
 
-        println!("Dropped database '{database_name}'");
+            println!("Dropped database '{database_name}'");
+        }
 
         Ok(())
     }
@@ -107,6 +118,7 @@ hillside15,2023-03-25T14:00:00Z,14.5,32.5,55.9,212.5
     #[cfg(test)]
     #[tokio::test]
     async fn load_data_twice() -> Result<(), Box<dyn std::error::Error>> {
+        let keep_database = env::var("DROP_DATABASE").unwrap_or(String::from("1")).eq("0");
         let data_source_name = "hub";
         let database_name = generate_name();
         let db_config = get_db_config()?;
@@ -133,39 +145,68 @@ hillside15,2023-03-25T14:00:00Z,14.5,32.5,55.9,212.5
         let mut cmd = Command::cargo_bin("minerva-admin")?;
         cmd.env("PGDATABASE", &database_name);
 
-        let mut csv_file = tempfile::tempfile().unwrap();
-
+        let mut csv_file = tempfile::NamedTempFile::new().unwrap();
         csv_file.write_all(TEST_CSV_DATA.as_bytes()).unwrap();
-
-        let instance_root_path = std::fs::canonicalize("../../examples/tiny_instance_v1").unwrap();
-
-        let mut file_path = PathBuf::from(instance_root_path);
-        file_path.push("sample-data/sample.csv");
 
         cmd.arg("load-data")
             .arg("--data-source")
             .arg(&data_source_name)
-            .arg(&file_path);
-        cmd.assert()
-            .success()
-            .stdout(predicate::str::contains("Job ID"));
+            .arg(&csv_file.path());
+
+        let output = cmd.output().unwrap();
+
+        println!("{}", String::from_utf8(output.stdout).unwrap());
 
         let mut cmd = Command::cargo_bin("minerva-admin")?;
         cmd.env("PGDATABASE", &database_name);
 
+        let mut csv_file = tempfile::NamedTempFile::new().unwrap();
+        csv_file.write_all(TEST_CSV_DATA_UPDATE_PARTIAL.as_bytes()).unwrap();
+
         cmd.arg("load-data")
             .arg("--data-source")
             .arg(&data_source_name)
-            .arg(&file_path);
-        cmd.assert()
-            .success()
-            .stdout(predicate::str::contains("Job ID"));
+            .arg(&csv_file.path());
 
-        let mut client = connect_to_db(&db_config).await?;
+        let output = cmd.output().unwrap();
 
-        drop_database(&mut client, &database_name).await?;
+        println!("{}", String::from_utf8(output.stdout).unwrap());
 
-        println!("Dropped database '{database_name}'");
+        let mut db_config_minerva = db_config.clone();
+        db_config_minerva.dbname(&database_name);
+
+        {
+            let client = connect_to_db(&db_config_minerva).await?;
+
+            let query = concat!(
+                "SELECT freq_power ",
+                "FROM trend.hub_node_main_15m t ",
+                "JOIN entity.node e ON e.id = t.entity_id ",
+                "WHERE e.name = $1 AND t.timestamp = $2::text::timestamptz"
+            );
+
+            let row = client.query_one(query, &[&"hillside14", &"2023-03-25T14:00:00Z"]).await?;
+
+            let expected_value: Decimal = dec!(212.4);
+            let value: Decimal = row.get(0);
+
+            assert_eq!(value, expected_value);
+
+            let row = client.query_one(query, &[&"hillside15", &"2023-03-25T14:00:00Z"]).await?;
+
+            let expected_value: Decimal = dec!(200.0);
+            let value: Decimal = row.get(0);
+
+            assert_eq!(value, expected_value);
+        }
+
+        if !keep_database {
+            let mut client = connect_to_db(&db_config).await?;
+
+            drop_database(&mut client, &database_name).await?;
+
+            println!("Dropped database '{database_name}'");
+        }
 
         Ok(())
     }
